requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
django-filter==23.5
django-cors-headers==4.3.1
python-decouple==3.8
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
Faker==22.0.0
coverage==7.4.0
